Add async to server and test

+++ server/server.js
@@ -15,65 +15,58 @@ const port = process.env.PORT || 3000;
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req, res) => {
-  var todo = new Todo({
+app.post('/todos', authenticate, async (req, res) => {
+  let todo = new Todo({
     text: req.body.text,
     _creator: req.user._id
   });
 
-  todo.save().then((doc) => {
+  try {
+    const doc = await todo.save();
     res.send(doc);
-  }).catch((e) => {
-    res.status(400).send(e);
-  });
+  } catch (e) {
+    res.status(400).send(e);  
+  }
 })
 
-app.get('/todos', authenticate, (req, res) => {
-  Todo.find({
-    _creator: req.user._id
-  }).then((todos) => {
-    res.send({
-      todos
-    })
-  }).catch((e) => {
-    res.status(400).send(e);
-  });
+app.get('/todos', authenticate, async (req, res) => {
+  try {
+    const todos = await Todo.find({ _creator: req.user._id });
+    res.send({ todos });
+  } catch (e) {
+    res.status(400).send(e);  
+  }
 });
 
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos/:id', authenticate, async (req, res) => {
   let id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
 
-  Todo.findOne({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+  try {
+    const todo = await Todo.findOne({ _id: id, _creator: req.user._id });  
     if (!todo) {
       return res.status(404).send();      
     } else {
-      res.status(200).send({
-        todo
-      });
+      res.status(200).send({ todo });
     }
-  }).catch((e) => {
-    res.status(400).send();
-  });
+  } catch (e) {
+    res.status(400).send();  
+  }
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
   let id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
 
-  Todo.findOneAndDelete({
-    _id: id,
-    _creator: req.user.id
-  }).then((todo) => {
+  try {
+    const todo = await Todo.findOneAndDelete({ _id: id, _creator: req.user.id });
+
     if (!todo) {
       return res.status(404).send('Not Found');      
     } else {
@@ -81,12 +74,12 @@ app.delete('/todos/:id', authenticate, (req, res) => {
         todo
       });
     }
-  }).catch((e) => {
-    res.status(400).send();
-  });
+  } catch (e) {
+    res.status(400).send();  
+  }
 });
 
-app.put(`/todos/:id`, authenticate, (req, res) => {
+app.put(`/todos/:id`, authenticate, async (req, res) => {
   let id = req.params.id;
   let body = _.pick(req.body, ['text', 'completed']);
 
@@ -101,32 +94,30 @@ app.put(`/todos/:id`, authenticate, (req, res) => {
     body.completedAt = null;
   }
 
-  Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, { $set: body }, {new: true}).then((todo) => {
+  try {
+    const todo = await Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, { $set: body }, {new: true});
     if (!todo) {
       return res.status(400).send();
     }
 
-    res.status(200).send({
-      todo
-    });
-  }).catch((e) => {
-    res.status(400).send();
-  });
+    res.status(200).send({ todo });
+  } catch (e) {
+    res.status(400).send();  
+  }
 });
 
 
 // POST /users
-app.post('/users', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-  var user = new User(body);
-
-  user.save().then(() => {
-    return user.generateAuthToken();
-  }).then((token) => {
-    res.header('x-auth', token).send(user);
-  }).catch((e) => {
-    res.status(400).send(e);
-  })
+app.post('/users', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = new User(body);
+    await user.save();
+    const token = user.generateAuthToken();
+    res.header('x-auth', token).send(user);  
+  } catch (e) {
+    res.status(400).send(e);  
+  }
 });
 
 app.get(`/users/me`, authenticate, (req, res) => {
@@ -134,28 +125,27 @@ app.get(`/users/me`, authenticate, (req, res) => {
 })
 
 // POST /users/login
-app.post(`/users/login`, (req, res) => {
-  let body = _.pick(req.body, ['email', 'password']);
-
-  User.findByCredentials(body.email, body.password).then((user) => {
-    return user.generateAuthToken().then((token) => { 
-      res.header('x-auth', token).send(user);
-    });
-  }).catch((e) => {
-    res.status(400).send();
-  });
+app.post(`/users/login`, async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);  
+  } catch (e) {
+    res.status(400).send();  
+  }
 }); 
 
-app.delete(`/users/me/token`, authenticate, (req, res) => {
-  let user = new User(req.user);
+app.delete(`/users/me/token`, authenticate, async (req, res) => {
+  try {
+    let user = new User(req.user);
 
-  user.removeToken(req.token).then(() => {
+    await user.removeToken(req.token);
     res.status(200).send();
-  }).catch((e) => {
-    res.status(400).send();
-  })
+  } catch (e) {
+    res.status(400).send();  
+  }
 })
-
 
 app.listen(port, () => {
   console.log(`Started up at port ${port} \n\n\n`)
